[fix/style] Fixed a bug with the international category and updated views to be more consistent.
There was a bug with the international categories which caused a null pointer
exception. This was fixed by saving a boolean which tells whether to show/add
the international category section to the list of sections, and making modifications
to the helper classes to account for this. The view width has also increased
in every view to account for the international category when displaying the
clue grid.